Make sure ccoctl binary is downloaded if doesn't exist in destroy job
Fixes: #9183

+++ ocs_ci/deployment/ibmcloud.py
@@ -196,6 +196,8 @@ class IBMCloudIPI(CloudDeploymentBase):
             logger.warning(
                 "Resource group for the cluster doesn't exist! Will not run installer to destroy the cluster!"
             )
+        # Make sure ccoctl is downloaded before using it in destroy job.
+        self.configure_cloud_credential_operator()
         self.delete_service_id()
         if resource_group:
             resource_group = self.get_resource_group()
@@ -472,17 +474,20 @@ class IBMCloudIPI(CloudDeploymentBase):
         the cluster while in manual mode.
 
         """
-        # retrieve ccoctl binary from https://mirror.openshift.com
-        version = config.DEPLOYMENT.get("ccoctl_version")
-        source = f"https://mirror.openshift.com/pub/openshift-v4/clients/ocp/{version}/ccoctl-linux.tar.gz"
         bin_dir = config.RUN["bin_dir"]
-        tarball = os.path.join(bin_dir, "ccoctl-linux.tar.gz")
-        logger.info("Downloading ccoctl tarball from %s", source)
-        download_file(source, tarball)
-        cmd = f"tar -xzC {bin_dir} -f {tarball} ccoctl"
-        logger.info("Extracting ccoctl binary from %s", tarball)
-        exec_cmd(cmd)
-        delete_file(tarball)
+        ccoctl_path = os.path.join(bin_dir, "ccoctl")
+        if not os.path.isfile(ccoctl_path):
+            # retrieve ccoctl binary from https://mirror.openshift.com
+            version = config.DEPLOYMENT.get("ccoctl_version")
+            source = f"https://mirror.openshift.com/pub/openshift-v4/clients/ocp/{version}/ccoctl-linux.tar.gz"
+            bin_dir = config.RUN["bin_dir"]
+            tarball = os.path.join(bin_dir, "ccoctl-linux.tar.gz")
+            logger.info("Downloading ccoctl tarball from %s", source)
+            download_file(source, tarball)
+            cmd = f"tar -xzC {bin_dir} -f {tarball} ccoctl"
+            logger.info("Extracting ccoctl binary from %s", tarball)
+            exec_cmd(cmd)
+            delete_file(tarball)
 
     @staticmethod
     def export_api_key():
